pyproject.toml
[tool.poetry]
name = "interview-agents"
version = "0.1.0"
description = "Interview agents using Google's ADK and Vertex AI"
authors = ["aminehd <aminehd@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "interview_agents"}]

[tool.poetry.dependencies]
python = ">=3.12"
requests = "^2.31.0"
google-adk = "^0.1.0"
pydantic = "^2.11.3"
python-dotenv = "^1.1.0"
google-cloud-aiplatform = {extras = ["adk", "agent_engines"], version = "^1.42.1"}
absl-py = "^2.1.0"
cloudpickle = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
